fix: moves certs code into ssl module
