+++ ssl/certs.go
@@ -7,7 +7,7 @@
 
 // Small modifications by kabukky
 
-package core
+package ssl
 
 import (
 	"crypto/ecdsa"
@@ -17,12 +17,13 @@ import (
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"encoding/pem"
-	"fmt"
 	"math/big"
 	"net"
 	"os"
 	"strings"
 	"time"
+
+	"github.com/op/go-logging"
 )
 
 var (
@@ -32,6 +33,8 @@ var (
 	rsaBits    = 2048
 	ecdsaCurve = ""
 )
+
+var log = logging.MustGetLogger("core")
 
 func publicKey(priv interface{}) interface{} {
 	switch k := priv.(type) {
@@ -51,7 +54,7 @@ func pemBlockForKey(priv interface{}) *pem.Block {
 	case *ecdsa.PrivateKey:
 		b, err := x509.MarshalECPrivateKey(k)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "Unable to marshal ECDSA private key: %v", err)
+			log.Errorf("Unable to marshal ECDSA private key: %s", err)
 			os.Exit(2)
 		}
 		return &pem.Block{Type: "EC PRIVATE KEY", Bytes: b}
@@ -84,11 +87,11 @@ func Generate(certPath string, keyPath string, host string) error {
 	case "P521":
 		priv, err = ecdsa.GenerateKey(elliptic.P521(), rand.Reader)
 	default:
-		fmt.Fprintf(os.Stderr, "Unrecognized elliptic curve: %q", ecdsaCurve)
+		log.Errorf("Unrecognized elliptic curve: %s", ecdsaCurve)
 		os.Exit(1)
 	}
 	if err != nil {
-		fmt.Printf("failed to generate private key: %s", err)
+		log.Errorf("failed to generate private key: %s", err)
 		return err
 	}
 
@@ -98,7 +101,7 @@ func Generate(certPath string, keyPath string, host string) error {
 	} else {
 		notBefore, err = time.Parse("Jan 2 15:04:05 2006", validFrom)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "Failed to parse creation date: %s\n", err)
+			log.Errorf("Failed to parse creation date: %s\n", err)
 			return err
 		}
 	}
@@ -108,7 +111,7 @@ func Generate(certPath string, keyPath string, host string) error {
 	serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
 	serialNumber, err := rand.Int(rand.Reader, serialNumberLimit)
 	if err != nil {
-		fmt.Printf("failed to generate serial number: %s", err)
+		log.Errorf("failed to generate serial number: %s", err)
 		return err
 	}
 
@@ -141,26 +144,26 @@ func Generate(certPath string, keyPath string, host string) error {
 
 	derBytes, err := x509.CreateCertificate(rand.Reader, &template, &template, publicKey(priv), priv)
 	if err != nil {
-		fmt.Printf("Failed to create certificate: %s", err)
+		log.Errorf("Failed to create certificate: %s", err)
 		return err
 	}
 
 	certOut, err := os.Create(certPath)
 	if err != nil {
-		fmt.Printf("failed to open "+certPath+" for writing: %s", err)
+		log.Errorf("failed to open "+certPath+" for writing: %s", err)
 		return err
 	}
 	pem.Encode(certOut, &pem.Block{Type: "CERTIFICATE", Bytes: derBytes})
 	certOut.Close()
-	fmt.Print("written cert.pem\n")
+	log.Infof("written cert.pem\n")
 
 	keyOut, err := os.OpenFile(keyPath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
 	if err != nil {
-		fmt.Print("failed to open "+keyPath+" for writing:", err)
+		log.Errorf("failed to open "+keyPath+" for writing:", err)
 		return err
 	}
 	pem.Encode(keyOut, pemBlockForKey(priv))
 	keyOut.Close()
-	fmt.Print("written key.pem\n")
+	log.Infof("written key.pem\n")
 	return nil
 }
